Separate dotplot, matrixplot and stackec_violin into individual files. Added separate file for `_baseplot_class.py` containing the BasePlot parent class.
